gen/check test process
