+++ .github/workflows/test.yml
@@ -18,7 +18,7 @@ concurrency:
 #   FORCE_COLOR: "1"
 
 jobs:
-  run:
+  matrix-test:
     name: Python ${{ matrix.python-version }} on ${{ startsWith(matrix.os, 'macos-') && 'macOS' || startsWith(matrix.os, 'windows-') && 'Windows' || 'Linux' }}
     runs-on: ${{ matrix.os }}
     strategy:
@@ -43,3 +43,71 @@ jobs:
 
     - name: Run tests
       run: hatch run test:test
+
+  patch:
+    name: Python ${{ matrix.python-version }} on ${{ startsWith(matrix.os, 'macos-') && 'macOS' || startsWith(matrix.os, 'windows-') && 'Windows' || 'Linux' }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: ['3.11']
+
+    steps:
+    - uses: actions/checkout@v3
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Ensure latest pip
+      run: python -m pip install --upgrade pip
+
+    - name: Install Hatch
+      run: python -m pip install hatch
+
+    - name: Run tests
+      env:
+        TESTWHEEL_GENERATE_PATH: ${{ runner.temp }}/testwheel-out/${{ matrix.python-version }}/${{ matrix.os }}
+      run: hatch run test:generate
+
+    - uses: actions/upload-artifact@v3
+      with:
+        name: patched-testwheel
+        path: ${{ runner.temp }}/testwheel-out
+
+  check:
+    name: Python ${{ matrix.python-version }} on ${{ startsWith(matrix.os, 'macos-') && 'macOS' || startsWith(matrix.os, 'windows-') && 'Windows' || 'Linux' }}
+    needs: patch
+
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: ['3.11']
+
+    steps:
+    - uses: actions/checkout@v3
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Ensure latest pip
+      run: python -m pip install --upgrade pip
+
+    - name: Install Hatch
+      run: python -m pip install hatch
+
+    - uses: actions/download-artifact@v3
+      with:
+        name: patched-testwheel
+        path: ${{ runner.temp }}/testwheel-out
+
+    - name: Run tests
+      env:
+        TESTWHEEL_GENERATE_PATH: ${{ runner.temp }}/testwheel-out/
+      run: hatch run test:check
