+++ tests/gen_check/test_generate.py
@@ -1,0 +1,15 @@
+import os
+import shutil
+from pathlib import Path
+
+import pytest
+
+
+if "TESTWHEEL_GENERATE_PATH" not in os.environ:
+    pytest.skip(allow_module_level=True)
+
+
+def test_store_patched_wheel(patched_wheel: Path) -> None:
+    out_dir = Path(os.environ["TESTWHEEL_GENERATE_PATH"])
+    out_file = out_dir / patched_wheel.name
+    shutil.copy(patched_wheel, out_file)
